+++ .vscode/settings.json
@@ -1,0 +1,5 @@
+{
+    "[json]": {
+        "editor.defaultFormatter": "esbenp.prettier-vscode"
+    }
+}
